Launcher program for the techchallenge.
The purpose of this launcher is to minimize the impact on the way
rectpack handles input and output while complying with the challenge's
requirements.

Since retpack has not been modified yet, the id and one coordinate of
the cuboids is ignored, and we are just printing all of rectpack's
output for now.